// src/enums/appointment.enum.ts
enum SYMPTOMS {
    TEMPERATURE = 'TEMPERATURE',
    DRY_COUGH = 'DRY_COUGH',
    BREATHING_PROBLEM = 'BREATHING_PROBLEM',
    BODY_PAIN = 'BODY_PAIN'
}

enum TEST_NAME_TYPE {
    EXPRESS = 'EXPRESS',
    PREMIUM = 'PREMIUM',
    VIP = 'VIP'
}

enum CUSTOMER_TYPE {
    WALK_IN = 'WALK_IN',
    HOME_SAMPLE = 'HOME_SAMPLE'
}

enum TEST_NAMES {
    BLOOD_TEST_URINE = 'BLOOD_TEST_URINE',
    COVID_19 = 'COVID_19',
    ULTRASOUND = 'ULTRASOUND',
    X_RAY = 'X_RAY',
    OTHERS = 'OTHERS'

}

enum CATEGORY {
    COVID_19 = 'COVID_19',
    GENERAL = 'GENERAL'
}

enum TYPE {
  NORMAL = "NORMAL",
  COVID_TEST = "COVID_TEST",
  COVID_VACINATIONd = 'COVID_VACINATION',
  DOCTOR_APPOINTMENTd = 'DOCTOR_APPOINTMENT',
  LAB_TEST = "LAB_TEST"
}

enum STATUS {
    BOOKED = 'BOOKED',
    EN_ROUTE = 'EN_ROUTE',
    ARRIVED = 'ARRIVED',
    CANCELLED = 'CANCELLED',
    SAMPLE_COLLECTED = 'SAMPLE_COLLECTED',
    COMPLETED = 'COMPLETED',
    PENDING = 'PENDING',
    SENT_TO_LAB = 'SENT_TO_LAB',
    BATCH_DONE = 'BATCH_DONE'
}

enum CAUSES {

    TRAVELLING = 'TRAVELLING',
    SCHOOL_REQUIRMENT = 'SCHOOL_REQUIRMENT',
    WORKPLACE_REQUIRMENT = 'WORKPLACE_REQUIRMENT',
    NOT_FEELING_WELL = "NOT_FEELING_WELL",
    OTHER = "OTHER"

}

enum APPOINTMENT_FOR {
    SELF = "SELF",
    FAMILY_MEMBER = "FAMILY_MEMBER"

}

enum PAYMENT_TYPE {
    CASH = "CASH", // Rider
    CARD = "CARD", // Vendor
    CREDIT_CARD = "CREDIT_CARD", // Bank
    ONLINE_TRANSFER = "ONLINE_TRANSFER", // Bank,

}

enum PAYMENT_STATUS {
    PAID = "PAID",
    PAY_AT_LOCATION = "PAY_AT_LOCATION"
}

enum HASANA_ID_STATUS {
    NOT_ENTERED = "NOT_ENTERED",
    ENTERED = "ENTERED"
}

enum TEAM_ASSIGN {
    ASSIGNED = "ASSIGNED",
    NOT_ASSIGNED = "NOT_ASSIGNED"
}

enum FILTER_TYPE {
    EID_PASSPORT = "EID_PASSPORT",
    VENDOR_ID_LAB_ID = "VENDOR_ID_LAB_ID",
    NAME = "NAME",
    EMAIL = "EMAIL",
    PHONE = "PHONE",
}
enum SENDTOLAB {
    NOT_SENDTOLAB = "NOT_SENDTOLAB",
    SEND_TO_LAB = "SEND_TO_LAB"
}

enum INITIAL_PAYMENT_TYPE {
    CARD = "CARD",
    PAID = "PAID",
    UN_PAID = "UN_PAID",
    CASH = "CASH", // Rider
    CREDIT_CARD = "CREDIT_CARD", // Bank
    ONLINE_TRANSFER = "ONLINE_TRANSFER", // Bank

}

export {
    SYMPTOMS,
    TEST_NAME_TYPE,
    CUSTOMER_TYPE,
    TEST_NAMES,
    CATEGORY,
    TYPE,
    STATUS,
    CAUSES,
    APPOINTMENT_FOR,
    PAYMENT_TYPE,
    PAYMENT_STATUS,
    HASANA_ID_STATUS,
    FILTER_TYPE,
    SENDTOLAB,
    TEAM_ASSIGN,
    INITIAL_PAYMENT_TYPE
}